rename _wrapper to _random_forest
