+++ external/fv3fit/fv3fit/sklearn/__init__.py
@@ -1,1 +1,1 @@
-from ._wrapper import SklearnWrapper, RegressorEnsemble
+from . import _random_forest
